Use Chromium Git mirror for AppRTC
AppRTC and the other WebRTC samples have moved to GitHub.
In crbug.com/377394 a mirror was setup for it that we can
use for the automated tests we have.

These modifications will make the change transparent to
the test (and hopefully it still passes with the latest
version).

BUG=362483
TEST=Locally syncing and running hooks and verifying the out/apprtc
contains the newly copied files.

Review URL: https://codereview.chromium.org/316013004

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/webrtc/webrtc.DEPS@275043 4ff67af0-8c30-449e-8e8b-ad334ec8d88c