Now starting AppRTC server up in "developer mode".
We don't have a way to click on the new join confirmation buttons from
Firefox; therefore we'll start the server in a special mode where the
join buttons are disabled.

We can't do this with an URL parameter since the purpose of the buttons
is to prevent embedding services to get users to auto-join.

B=None
R=kjellander@chromium.org

Review URL: https://codereview.chromium.org/872143005

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/webrtc/webrtc.DEPS@293859 4ff67af0-8c30-449e-8e8b-ad334ec8d88c