Make Firefox nightly download more robust.
The script will now only start complaining if we haven't been able to
download anything for days. The script will always touch the file if
the download is successful, so the modified time of the latest
downloaded time is the benchmark of how fresh our Firefox build is.
There can be many days between FF nightly release, sometimes weeks,
so we only care if we managed to download something and not which date
the build is from.

BUG=360516
R=kjellander@chromium.org

Review URL: https://codereview.chromium.org/883273002

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/webrtc/webrtc.DEPS@293895 4ff67af0-8c30-449e-8e8b-ad334ec8d88c