Now cleaning old firefox archives.
We have a few gigs of old archives out on the bots, and nothing
to clean them up. There's no reason to keep more than one old
archive around, so this makes the download script prune them.

I considered making this a separate script, but arguably it's the
job of the download script to deal with its byproducts (and I
could share some code) so I made the existing script do it.

BUG=360516
R=kjellander@chromium.org

Review URL: https://codereview.chromium.org/891303002

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/webrtc/webrtc.DEPS@293897 4ff67af0-8c30-449e-8e8b-ad334ec8d88c