Modify app.yaml in the built AppRTC rather in src.
When I switched to prebuilt AppRTC, I forgot we need to modify the
built app.yaml rather the one in src. This used to work because we ran
grunt build after the copy (which copied the file I modified), but after
switching to pre-built we're effectively running grunt build _before_
the copy. This caused copy_apprtc.py to write to a config file that
wasn't read by the server on startup.

We could also delete the code and just flip the flag during apprtc.zip
packaging, but that seems tedious and error-prone.

BUG=581283
R=kjellander@chromium.org

Review URL: https://codereview.chromium.org/1657693002 .